create testing for extra points class
