+++ app/src/main/java/pl/motyliar/budgetapp/domain/character/subclass/ExtraPoints.kt
@@ -17,6 +17,7 @@ class ExtraPoints {
     }
 
     fun setExtraPoints(value: Int) {
+        if(value > 0)
         extraPoints += value
     }
